[feat] Add ItemDB
아이템의 ID를 기준으로 정리된 데이터 베이스를 만들어서 SO로 유지하는 방식 구현

+++ Assets/02.Scripts/Player/PlayerLevelSystem.cs
@@ -2,12 +2,12 @@
 
 public class PlayerLevelSystem
 {
-    public int InGameCurrentExp { get; private set; }
-    public int InGameMaxExp { get; private set; }
+    public float InGameCurrentExp { get; private set; }
+    public float InGameMaxExp { get; private set; }
     public int InGameLevel { get; private set; }
 
-    public int OutGameCurrentExp { get; private set; }
-    public int OutGameMaxExp { get; private set; }
+    public float OutGameCurrentExp { get; private set; }
+    public float OutGameMaxExp { get; private set; }
     public int OutGameLevel { get; private set; }
